優化

+++ src/main/java/cn/tedu/mall/service/impl/UserAddressServiceImpl.java
@@ -2,8 +2,10 @@ package cn.tedu.mall.service.impl;
 
 import cn.tedu.mall.exception.ServiceException;
 import cn.tedu.mall.mapper.UserAddressMapper;
+import cn.tedu.mall.mapper.UserMapper;
 import cn.tedu.mall.pojo.domain.Address;
 import cn.tedu.mall.pojo.user.UserAddressDTO;
+import cn.tedu.mall.pojo.user.UserUpdateDTO;
 import cn.tedu.mall.service.IUserAddressService;
 import cn.tedu.mall.utils.ConstUtils;
 import cn.tedu.mall.web.ServiceCode;
@@ -26,12 +28,24 @@ public class UserAddressServiceImpl implements IUserAddressService {
     @Autowired
     private UserAddressMapper userAddressMapper;
 
+    @Autowired
+    private UserMapper userMapper;
+
+    public static final int MAX_ADDRESS_COUNT = 5;
 
     @Override
     public void insert(UserAddressDTO userAddressDTO) {
         log.debug("開始新增地址");
         //獲取userId
         Long userId = ConstUtils.getUserId();
+        //根據用戶id獲取目前有多少地址
+        int count = userAddressMapper.updateAddressDefaultByUserId(userId);
+        //判斷儲存的地址數量是否超過最大值，目前設定為5
+        if(count >=MAX_ADDRESS_COUNT){
+            //超過，拋異常
+            throw new ServiceException(ServiceCode.ERR_BAD_REQUEST,"儲存量已達最大");
+        }
+
         userAddressDTO.setUserId(userId);
         //判斷該地址是否是預設
         if(userAddressDTO.getIsDefault() == ConstUtils.IS_DEFAULT){
@@ -41,6 +55,13 @@ public class UserAddressServiceImpl implements IUserAddressService {
         int rows = userAddressMapper.insertAddress(userAddressDTO);
         if (rows !=1){
             throw new ServiceException(ServiceCode.ERR_INSERT,"伺服器忙碌請稍候!");
+        }
+        UserUpdateDTO userUpdateDTO = new UserUpdateDTO();
+        userUpdateDTO.setDefaultAddId(userAddressDTO.getId());
+        userUpdateDTO.setId(userId);
+        int updateRow = userMapper.update(userUpdateDTO);
+        if (updateRow !=1){
+            throw new ServiceException(ServiceCode.ERR_UPDATE,"伺服器忙碌請稍候!");
         }
     }
 
@@ -97,6 +118,14 @@ public class UserAddressServiceImpl implements IUserAddressService {
     @Override
     public void deleteById(Long id) {
         log.debug("開始根據id刪除地址");
+        //獲取地址資料
+        Address getAddVO = userAddressMapper.getAddressById(id);
+        //判斷要刪除的地址是否為預設
+        if(getAddVO.getIsDefault() == ConstUtils.IS_DEFAULT){
+            //是預設，拋異常
+            throw new ServiceException(ServiceCode.ERR_DELETE,"不能刪除預設地址!");
+        }
+
         int rows = userAddressMapper.deleteById(id);
         if (rows !=1){
             throw new ServiceException(ServiceCode.ERR_DELETE,"伺服器忙碌請稍候!");
